+++ number/gcd.hpp
@@ -1,6 +1,6 @@
 /**
  * An implementation of the greatest common divisor finding algorithm in C++.
- * The algorithm supports natural numbers. It is provided in two variants:
+ * The algorithm supports natural number. It is provided in two variants:
  * the standard one (euclidean) and the one based on Stein's algorithm
  * that is without division or modulo operations which are typically significantly
  * slower than the logic operations.
@@ -10,20 +10,22 @@
 #ifndef ALGORITHMS_GCD_H
 #define ALGORITHMS_GCD_H
 
+#include <cstdint>
+
 /**
  * Find GCD using the extended euclidean algorithm.
  * @param x, y  arguments
  * @param s, t  coefficients such that GCD(x, y) = xs + yt
  * @return      GCD(x, y)
  */
-int gcd(int x, int y, int &s, int &t) {
+int64_t gcd(int64_t x, int64_t y, int64_t &s, int64_t &t) {
     if (y == 0) {
         s = 1;
         t = 0;
         return x;
     }
-    int s1, t1;
-    int d = gcd(y, x % y, s1, t1);
+    int64_t s1, t1;
+    int64_t d = gcd(y, x % y, s1, t1);
     s = t1;
     t = s1 - (x / y) * t1;
     return d;
@@ -35,8 +37,8 @@ int gcd(int x, int y, int &s, int &t) {
  * @param y         second number
  * @return          gcd(x, y)
  */
-int gcd_stein(int x, int y) {
-    int shift_counter = 0;
+int64_t gcd_stein(int64_t x, int64_t y) {
+    int64_t shift_counter = 0;
     while (y > 0 && x > 0) {
         if (1 - (x & 1) && 1 - (y & 1)) {
             x = x >> 1;
@@ -49,7 +51,7 @@ int gcd_stein(int x, int y) {
         } else if (x >= y) {
             x = (x - y) >> 1;
         } else {
-            int z = x;
+            int64_t z = x;
             x = y;
             y = z;
             x = (x - y) >> 1;
